Revert "Allow crosshair size below 1 (0.25/0.5/0.75)"
This reverts commit 5ea5581a12db33e2b94351f4d3d7276f8acadd77.

+++ mod/src/main/java/com/tntsallin1client/config/ClientConfig.java
@@ -68,9 +68,7 @@ public class ClientConfig {
 	public CrosshairMode crosshairMode = CrosshairMode.PRESET;
 	public CrosshairPreset crosshairPreset = CrosshairPreset.SMALLER;
 	public boolean[][] crosshairCustomGrid = CrosshairGrid.empty();
-	// Float, not int: the size list has non-uniform steps (0.25/0.5/0.75 below
-	// 1, then whole numbers 1-6 above), see DiscreteSliderButton/CrosshairOptionsScreen.
-	public float crosshairPixelSize = 2.0F;
+	public int crosshairPixelSize = 2;
 	public boolean crosshairIgnoreGuiScale = false;
 	public boolean crosshairTargetColorEnabled = false;
 	public int crosshairTargetColor = 0xFFFF5555;
